Added version check for column drop

+++ app/src/main/kotlin/at/bitfire/davdroid/db/AppDatabase.kt
@@ -8,8 +8,8 @@ import android.accounts.AccountManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.database.sqlite.SQLiteException
 import android.database.sqlite.SQLiteQueryBuilder
+import android.os.Build
 import androidx.annotation.VisibleForTesting
 import androidx.core.app.NotificationCompat
 import androidx.core.database.getStringOrNull
@@ -138,11 +138,10 @@ abstract class AppDatabase: RoomDatabase() {
                             db.execSQL("UPDATE collection SET timezoneId=? WHERE id=?", arrayOf(timezoneId, id))
                         }
                     }
-                    try {
+                    // DROP COLUMN is available since SQLite 3.35.0, which is available since API 34,
+                    // in older versions just keep the timezone column
+                    if (Build.VERSION.SDK_INT > Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
                         db.execSQL("ALTER TABLE collection DROP COLUMN timezone")
-                    } catch (_: SQLiteException) {
-                        // DROP COLUMN is not available on old versions of SQLite, so if the drop
-                        // fails, just keep the "timezone" column.
                     }
                 }
             },
